requirements.txt
quart
quart-cors
python-pptx